typescript(vx-demo): rewrite StackedAreas in TypeScript
